requirements.txt
pandas
requests
python-dotenv
ipykernel
xgboost